made a working scale appender using a class!
