+++ programs/shortScripts/classCreation/functions/__init__.py
@@ -1,0 +1,2 @@
+from .dataRead import parseData, jsonRead
+
