+++ programs/shortScripts/classCreation/functions/data.json
@@ -1,0 +1,1 @@
+{"major": ["Major", "Minor", "Minor", "Major", "Major", "Minor", "Diminished"], "minor": ["Minor", "Diminished", "Major", "Minor", "Minor", "Major", "Major"], "dorian": ["Minor", "Minor", "Major", "Major", "Minor", "Diminished", "Major"], "phrygian": ["Minor", "Major", "Major", "Minor", "Diminished", "Major", "Minor"], "liydian": ["Major", "Major", "Minor", "Diminished", "Major", "Minor", "Minor"], "mixoliydian": ["Major", "Minor", "Diminished", "Major", "Minor", "Minor", "Major"], "locrian": ["Major", "Major", "Minor", "Diminished", "Major", "Minor", "Minor"], "majorIntervals": [2, 2, 1, 2, 2, 2, 1], "minorIntervals": [2, 1, 2, 2, 1, 2, 2], "dorianIntervals": [2, 1, 2, 2, 2, 1, 2], "phrygianIntervals": [1, 2, 2, 2, 1, 2, 2], "lydianIntervals": [2, 2, 2, 1, 2, 2, 1], "mixolydianIntervals": [2, 2, 1, 1, 2, 1, 2], "locrianIntervals": [1, 2, 2, 1, 2, 2, 2], "majorPentatonics": [2, 2, 3, 2, 3], "minorPentatonics": [3, 2, 2, 3, 2], "Sharps": ["C", "C#", "D", "D#", "E", "F", "F#", "G", "G#", "A", "A#", "B"], "Flats": ["C", "Db", "D", "Eb", "E", "F", "Gb", "G", "Ab", "A", "Bb", "B"]}
